Multiple fixes
Only attempt to delete the attached file to the email if one was sent as
part of the submit.

Comment out setting email message content as text/html, only standard
email body for now (text/plain).

+++ container_status_app/views/nas_notifications.py
@@ -201,8 +201,11 @@ class NasNotifications(MethodView):
             # email_sent = email.send_email(request.form.get('email_content'))
             email_sent = email.send_email(request.form.get('email_content'),
                                           attachment=file)
-            if os.path.exists(os.path.join(container_status_app.config.get('UPLOAD_FOLDER'), file.filename)):
-                os.remove(os.path.join(container_status_app.config.get('UPLOAD_FOLDER'), file.filename))
+
+            if request.files.get('attachment_file') is not None:
+                if os.path.exists(os.path.join(container_status_app.config.get('UPLOAD_FOLDER'), file.filename)):
+                    os.remove(os.path.join(container_status_app.config.get('UPLOAD_FOLDER'), file.filename))
+
             if email_sent:
                 container_status_app.logger.info(email_sent_message.format('sent'))
                 return True
